More stats
I used the picture cat.jpg to act as the user’s pet. I also tried to get my background to work and be bigger. Currently, when it is running it just turns into a black screen. I also added more stats to the pet and its health meters such as playfulness. I also looked for more transparent images to use for things like food, a bed, or a toy so that I can click on it and make the pet’s stats higher.