I added more sprites.
I got the screen to work and I downloaded the background image and cat food for the user to feed the cat with. I made health meters where it fluctuates depending on if you care for the cat or not. I’m having certain problems such as the background not being able to fit completely on the pygame screen. I am also trying to get my bar meters to show the percentage that it is at instead of looking at the bar and assuming how low/high it is. I made it so that I can drag the cat food to the cat to make the game more interactive. I also have the coin system implemented, where the user can only get food if they have enough coins.