resized background & buttons
I noticed that the background was too big for the screen size. The screen only showed a part of the background so I looked into how to resize the background. I also didn’t like how the buttons for interaction had a white/gray box around it, so I changed it to only words and no box.