Modified Status for cat
I changed the status that appears on screen so that it is labeled. Before this, the status would fluctuate but I wouldn’t know which status of the cat it's referring to. I’m having trouble making the cat and cat food have transparent backgrounds. I also tried to add another status for the cat’s thirst using the newly downloaded water.png. Something that I want to work on is making both the cat food and water appear on the bottom left corner of the screen. Even if I drag them to feed the cat, they should always respawn at the bottom left corner.