Revised sprites + mood meter
I looked at my code and thought about what I wanted to add to the code in the next few days. 
This is what I want to be added to my pet simulator:
Add an introduction screen with instructions and a ‘play’ button to click.
Make the sprite have a transparent background.
Add captions for when the sprite’s mood meters are dropping.
Add buttons for when the user wants to pause the game, go back to the introduction screen, or end the game.
Adjust my sprite sizes and add music!