Introduction screen Debug
I got stuck on a bug with my introduction screen. I’m trying to make a ‘play’ button which appears on the introduction screen so that when the player clicks it, the user can move onto the game and the introduction screen turns to False. I’m having trouble trying to transition from the introduction screen into the actual screen.