Added coin sprite
I looked at previous projects to remind myself how to use keys to control sprites. I want to be able to move my cat using WASD to allow the user to get coins if the cat overlaps with the coin sprite. I revised sizing of the cat sprite as well as the positioning of it.